refactor:  StatusSet를 getStatusSet 로 rename 후 getStatusLines과 rfc로 수정
