--- src/shared/components/promisisDetail/utils/StatusSet.tsx
@@ -1,66 +1,0 @@
-import {
-  IcVoteBlue,
-  IcVotingRed,
-  IcVotingYellow,
-  IcPinLocBlue,
-  IcPinLocYellow,
-  IcPinLocRed,
-  IcDateBlue,
-  IcDateYellow,
-  IcDateRed,
-} from '@svg/index.ts';
-import React from 'react';
-import type { PromisisDetailProps } from '@shared/components/promiseContainer/types/PromiseContainer.type';
-import { formatDateWithDay } from '@shared/utils/formatDateWithDay';
-import { PROMISE_STATUS, type PromiseStatusType } from '@shared/constant/promiseStatus';
-
-interface StatusIconSet {
-  vote: React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
-  pin: React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
-  date: React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
-}
-
-export const StatusSet: Record<PromiseStatusType, StatusIconSet> = {
-  [PROMISE_STATUS.PENDING]: {
-    vote: IcVoteBlue,
-    pin: IcPinLocBlue,
-    date: IcDateBlue,
-  },
-  [PROMISE_STATUS.VOTING]: {
-    vote: IcVotingYellow,
-    pin: IcPinLocYellow,
-    date: IcDateYellow,
-  },
-  [PROMISE_STATUS.CONFIRMED]: {
-    vote: IcVotingRed,
-    pin: IcPinLocRed,
-    date: IcDateRed,
-  },
-};
-
-export function getStatusLines(props: PromisisDetailProps): {
-  firstLine: string;
-  secondLine: string;
-  thirdLine: string;
-} {
-  switch (props.status) {
-    case PROMISE_STATUS.PENDING:
-      return {
-        firstLine: `${props.untilVoteDate}일`,
-        secondLine: props.suggestedRegion,
-        thirdLine: `${formatDateWithDay(props.suggestedStartDate)} ~ ${formatDateWithDay(props.suggestedEndDate)}`,
-      };
-    case PROMISE_STATUS.VOTING:
-      return {
-        firstLine: `${props.untilVoteEndDate}일`,
-        secondLine: props.suggestedRegion,
-        thirdLine: `${formatDateWithDay(props.suggestedStartDate)} ~ ${formatDateWithDay(props.suggestedEndDate)}`,
-      };
-    case PROMISE_STATUS.CONFIRMED:
-      return {
-        firstLine: props.promiseMembersNames.join(', '),
-        secondLine: props.placeName,
-        thirdLine: formatDateWithDay(props.fixedDate),
-      };
-  }
-}
